% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{clip_hobotemp}
\alias{clip_hobotemp}
\title{clip hobotemp table by date}
\usage{
clip_hobotemp(x, deploy = NA, recover = NA)
}
\arguments{
\item{x}{tibble, hobotemp}

\item{deploy}{POSIXt or NA, if not NA, clip data before this time}

\item{recover}{POSIXt or NA, if not NA, clip data before this time}
}
\value{
tibble
}
\description{
clip hobotemp table by date
}
