% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{draw_line_plot}
\alias{draw_line_plot}
\title{Plot temperature data as line plot, can facet to multiple sites}
\usage{
draw_line_plot(
  x = example_ridge_data(),
  main = "Temperature at Depth",
  xlabel = "Date",
  ylabel = "Temperature (degrees C)",
  ordered = NULL,
  facet = NULL,
  ...
)
}
\arguments{
\item{x}{tibble of hobotemp data}

\item{main}{character, title}

\item{xlabel}{character, title of xaxis}

\item{ylabel}{character, title of yaxis}

\item{ordered}{vector, vector of site names in order to be displayed bottom to top in legend and/or when facetted, must match site names in csv, default NULL}

\item{facet}{character, column to facet upon, default NULL}

\item{...}{further arguments passed to \code{\link[ggplot2]{theme}}}
}
\value{
ggplot2 object
}
\description{
Plot temperature data as line plot, can facet to multiple sites
}
