% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{draw_plot}
\alias{draw_plot}
\title{Plot temperature data as scatter plot with trendline}
\usage{
draw_plot(
  x = read_hobotemp(),
  main = "Temperature at Depth",
  xlabel = "Date",
  ylabel = "Temperature (degrees C)",
  alpha = 0.2,
  ...
)
}
\arguments{
\item{x}{tibble of hobotemp data}

\item{main}{character, title}

\item{xlabel}{character, title of xaxis}

\item{ylabel}{character, title of yaxis}

\item{alpha}{numeric, 0.1 default}

\item{...}{further arguments passed to \code{\link[ggplot2]{theme}}}
}
\value{
ggplot2 object
}
\description{
Plot temperature data as scatter plot with trendline
}
