% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{draw_ridgeline_plot}
\alias{draw_ridgeline_plot}
\title{Plot temperature data as ridgeline plot}
\usage{
draw_ridgeline_plot(
  x = example_ridge_data(),
  main = "Temperature at Depth",
  xlabel = "Date",
  ylabel = "Site",
  ordered = NULL,
  ...
)
}
\arguments{
\item{x}{tibble of hobotemp data with DateTime, Temp, and Site columns}

\item{main}{character, title}

\item{xlabel}{character, title of xaxis}

\item{ylabel}{character, title of yaxis}

\item{ordered}{vector, vector of site names in order to be displayed bottom to top, must match site names in csv}

\item{...}{further arguments passed to \code{\link[ggplot2]{theme}}}
}
\value{
ggplot2 object
}
\description{
Plot temperature data as ridgeline plot
}
