% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{example_ridge_data}
\alias{example_ridge_data}
\title{retrieve example type hobotemp csv with multi sites}
\usage{
example_ridge_data()
}
\value{
hobotemp csv
}
\description{
retrieve example type hobotemp csv with multi sites
}
