% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_hobo_cols}
\alias{read_hobo_cols}
\title{read hobotemp data file header}
\usage{
read_hobo_cols(filename = example_filename(), skip = 1)
}
\arguments{
\item{filename}{character, the name of the file}

\item{skip}{numeric, number of lines to skip - default 1}
}
\value{
named list
}
\description{
read hobotemp data file header
}
