% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{summarize_hobotemp}
\alias{summarize_hobotemp}
\title{print out summary of hobotemp data}
\usage{
summarize_hobotemp(x = read_hobotemp())
}
\arguments{
\item{x}{tibble, tibble of hobotemp data}
}
\value{
tibble
}
\description{
print out summary of hobotemp data
}
