% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{clip_hobotemp}
\alias{clip_hobotemp}
\title{clip hobotemp table by date}
\usage{
clip_hobotemp(x, startstop = NA)
}
\arguments{
\item{x}{tibble, hobotemp}

\item{startstop}{POSIXt vector of two values or NA, only used if clip = "user"}
}
\value{
tibble
}
\description{
clip hobotemp table by date
}
