% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_hobotemp}
\alias{read_hobotemp}
\title{read hobotemp data file}
\usage{
read_hobotemp(
  filename = example_filename(),
  clipped = c("auto", "user", "none")[1],
  startstop = NA
)
}
\arguments{
\item{filename}{character, the name of the file}

\item{clipped}{character, if auto, removed partial start/end days. if user, uses supplied startstop days. if none, does no date trimming}

\item{startstop}{POSIXt vector of two values or NA, only used if clip = "user"}
}
\value{
tibble
}
\description{
read hobotemp data file
}
